rename files to follow conventions
