+++ lib/index.js
@@ -1,2 +1,2 @@
 require("./foods-styling.scss");
-require("./foods-requests");
+require("./foods");
